% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdp_contrib.R
\name{ggplot_1_contrib}
\alias{ggplot_1_contrib}
\title{Title}
\usage{
ggplot_1_contrib(
  res_all,
  tgt_var,
  in_data,
  pps = FALSE,
  smooth = FALSE,
  title = NULL,
  x_units = "",
  y_units = "",
  n_x = 100,
  the_theme = theme_bw
)
}
\arguments{
\item{the_theme}{}
}
\description{
Title
}
