% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdp_contrib.R
\name{ggplot_1_roi}
\alias{ggplot_1_roi}
\title{Title}
\usage{
ggplot_1_roi(
  in_contribs,
  tgt_var,
  in_data = NULL,
  x_min = ifelse(!is.null(in_data), min(in_data[[tgt_var]]), x_min),
  x_max = ifelse(!is.null(in_data), max(in_data[[tgt_var]]), x_max),
  title = NULL,
  x_units = "",
  y_units = "",
  n_x = 100,
  the_theme = theme_bw
)
}
\arguments{
\item{the_theme}{}
}
\description{
Title
}
