% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdp_contrib.R
\name{pdp_contribs}
\alias{pdp_contribs}
\title{Title}
\usage{
pdp_contribs(
  in_model,
  in_data,
  pred_vars,
  pred_fun,
  grid_resolution = 20,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
