% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_contribs.R
\name{shap_contribs}
\alias{shap_contribs}
\title{Title}
\usage{
shap_contribs(
  in_model,
  in_data,
  pred_vars,
  pred_fun,
  nsim = 10,
  grid_resolution = 20,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
